pluginlib: Add per plugin IO timeout for backend.
Backend can define compile in timeout value with the following variable:
`const uint32_t BACKEND_TIMEOUT = 0; // use default`
The value set at this variable defines a timeout in seconds which will
be applied to all recvbackend_data() and sendbackend_data() calls.
The value of zero (0) setup a default timeout of 3600 seconds.
Set it to zero if unsure.