Fix #9224 about Storage Daemon issue with BoostFS
On bootfs, the call to ioctl(tmp_fd, FS_IOC_GETFLAGS, (char *)&get_attr);
triggers an error, and the get_attr variable is not long enough
and the memory in the stack get overwritten leading to a segfault.

The man page is incorrect, a "int" is documented, but a "long"
is actually needed.