Fix #9764 Abort a job if a PluginOptions is incorrect
If a plugin is not installed and is specified with the Job PluginOptions, the
job simply terminated "Verify OK"

With other plugins, like the vSphere plugin, if the plugin is not installed,
but a Fileset calls it, we get a failed job and a warning about "command not
found" or similar.

The important part here is that if the plugin is missing, we end up with False
negatives that a sysadmin should be notified about.